fix(serve,docs): a test that could not fail, and a test that was right about a code bug the spec argues for
Phase E begins. Two CodeRabbit findings reproduced against HEAD.

C1 — THE FIXTURE NEVER FINISHED THE REQUEST IT WAS TESTING.

a_chunked_body_past_the_cap_is_cut_off wrote a declared chunk size of 3 MiB and pushed the
payload, and never sent the chunk's trailing CRLF NOR the terminating zero-length chunk. So
the request was a truncated chunked body, not a well-formed one. Its only assertion was
"not a 200" — and a server that ignored max_request_bytes entirely would also produce no
200, because it would be waiting for bytes that never arrive. The test could not
distinguish "rejected for exceeding the cap" from "never completed", which is the only
distinction it exists to make. It passed for the wrong reason and would have kept passing
if the cap were deleted.

Fixed by sending \r\n and 0\r\n\r\n after the payload, keeping the ignored write errors (a
broken pipe mid-stream is the expected outcome) and the 200-rejection assertion.

Fault-injected by raising MAX_REQUEST_BYTES from 2 MiB to 64 MiB: the test now fails with
"a body over max_request_bytes must not be accepted". It could not have done that before —
the injection would have produced no response and the assertion would have passed.

C8 — THE FIXTURE WAS RIGHT; THE CODE WAS WRONG, AND THE SPEC SAYS SO.

The finding proposed fixing the PARSER: "join data_lines without popping a character so
trailing newlines survive". Adding the payloads reproduced a real failure at once:

    payload "a\n" was mangled — left: "a", right: "a\n"

But the proposed fix is wrong. The SSE specification requires the client to "remove the last
character from the data buffer if it is a U+000A LINE FEED" — the pop is mandated, and
removing it would make `data:a\ndata:b\n\n` parse as "a\nb\n" instead of "a\nb", breaking
every multi-line event. A finding is a lead, not an authority.

The bug was in the encoder. split_lines did
`s.split(['\n','\r']).filter(|l| !l.is_empty())`, so a payload ending in a line break lost
its trailing empty field AND the break. The fix is the opposite of the finding's: keep every
field, and emit ONE EXTRA `data:` line when the payload ends with a line break, so the
specification's mandatory trim restores it. data("a\n") is now three data lines, which join
to "a\n\n" and trim to "a\n".

Splitting on both terminators in one pass was also wrong for "a\r\nb" (three fields, one
spurious); the new splitter consumes \r\n as a single break.

THREE ARTEFACTS ASSERTED THE LOSSY BEHAVIOUR AS A REQUIREMENT, which is why a failing
round-trip never surfaced as a defect:

  1. the encoder's doc comment argued it was necessary — "the 'at least one' is what makes
     data(""), data("\n") and data("\n\n") encode to the same bytes, which they must". They
     must not: identical bytes cannot carry three different payloads, so the claim amounts to
     saying a trailing line break is not transmittable;
  2. split_lines' doc comment explained why empty fields are dropped and called the two rules
     "a pair";
  3. a unit test NAMED THE BUG AS A CONTRACT: a_newline_payload_is_the_same_event_as_an_empty_one.

A test that passes for the wrong reason usually has a comment, and often a test name,
asserting the wrong thing as a requirement. Changing the fixture alone would have made the
suite fail without saying which of the three was load-bearing; the reproduction is what
showed the parser was the one piece that was right. Replaced with
a_trailing_line_break_survives_the_encoder, which applies the specification's assembler rule
to the encoder's own output and requires the payload back.

Fault-injected by disarming the extra-line rule: the round-trip test panics. Restored: the
whole crate passes — 681 lib tests plus every integration suite, 0 failures.

A VERIFICATION GAP OF MY OWN, FROM ROUND 7.

cargo clippy on the changed crate reported unnecessary_cast (usize -> usize) in the
lifecycle.rs assertion added last round. I did not catch it then because round 7 ran
`cargo fmt --all -- --check` and the Python checker gate — and THE PYTHON GATE DOES NOT
COMPILE RUST. Definition of done #3 asks for both fmt and clippy; I ran one. Fixed, and the
lesson is that "the gate passed" is only as strong as the gate's coverage. fmt says nothing
about lints, and a lint is a new rule arriving with each Rust release — Phase C's subject.

Verified: cargo fmt --all -- --check exit 0; cargo clippy -p qqq-serve --all-targets
--all-features -- -D warnings exit 0; cargo test -p qqq-serve exit 0.

Recorded as §O-278.

+++ crates/qqq-serve/src/lifecycle.rs
@@ -1088,7 +1088,7 @@ mod tests {
         // step 10 to `Partial` made it need `thirteen`, and nothing would have said so.
         // A count in prose that nothing compares is exactly the defect this test exists for
         // (`§O-277`), so it is compared here too.
-        let not_done = STAGES.len() - Counts::measured().summary.implemented as usize;
+        let not_done = STAGES.len() - Counts::measured().summary.implemented;
         let narrative = format!("false for **{not_done}** of the fifteen");
         assert!(
             src.contains(&narrative),
